AHN point clouds area

Get AHN point clouds of a certain area.

AHN Point clouds is retrieved through a circle (x, y and radius), BBOX (x, y and radius, OR bbox), or own geometry (polygon). AHN data is obtained from the AHN3 (default), AHN2 or AHN 1 from the available point clouds datasets. AHN3 (default) only has one available datasets. AHN1 and AHN2 include 'gefilterd' or 'uitgefilterd' datasets. See documentation for all available parameters.

ahn_pc(
  name = "AHNpointclouds",
  output.dir = "AHN_output",
  X,
  Y,
  radius,
  bbox,
  polygon,
  AHN = "AHN3",
  gefilterd = FALSE,
  LONLAT = FALSE,
  sheets.location,
  sheets.keep = TRUE,
  sheets.redownload = FALSE
)

Arguments

name
Optional. Give a name of the specified area. This name will be used in the output file names.

output.dir
Set location of output directory. Default 'AHN_output'. 'AHN_output' saves all the output files in the 'AHN_output' directory created in the working directory. Set to any other output path to save all the output files there. Set to "", to have all output files be temporary files. This output directory excludes the location of the AHN sheets which is depicted with the 'sheets.location' parameter.

X
Required for circle or bbox when no BBOX coordinates are provided. X coordinate in RD New or WGS84 (LON)

Y
Required for circle or bbox when no BBOX coordinates are provided. Y coordinate in RD New or WGS84 (LAT)

radius
Required for circle or bbox when no BBOX coordinates are provided. Set radius in meters of area around a point to create a buffer area (circle).

bbox
Required when using a BBOX. Create BBOX of an area. Use c(XMIN, YMIN, XMAX, YMAX) OR set to TRUE when a radius is provided. Use RD New or WGS84 (LONLAT) coordinates.

polygon
Required when create an area of custom geometry. Use spatial object as your area. .shp, .gpkg. Output will in the shape of this polygon

AHN
Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'

gefilterd
Default FALSE. Only applicable for AHN1 or AHN2 point cloud data. It depicts if you want to download the 'gefilterd' point cloud data set or the 'uitgefilterd' data set.

LONLAT
Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will always be in RD New format

sheets.location
Default is the 'AHN_sheets' directory in the working directory. Set directory where all the AHN sheets are loaded when pre-exisiting sheets will be used or when new sheets will be stored. When loading existing files, always use the correct directory structure and capitalization within the selected directory. Example directory structure when this parameter is set to e.g. 'myFolder': 'myFolder/AHN_sheets/AHN3/PC' or 'myFolder/AHN_sheets/AHN2/PC'. Only use extracted files in their original name after download.

sheets.keep
Default TRUE. Set to FALSE if you want to delete the downloaded point clouds sheets (structure). It is recommended to keep the sheets if this function will be used more than once.

sheets.redownload
Default FALSE. Only applicable if point clouds sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (structure)

Value

.laz data of area