Download DTM

Download DTM download_dtm(name, output.dir, AHN = "AHN3", dem = "DSM", resolution = 0.5, bladnrs, area, interpolate = TRUE, sheets.keep = TRUE, sheets.redownload = FALSE)

download_dtm(
  name,
  output.dir,
  AHN = "AHN3",
  dem = "DSM",
  resolution = 0.5,
  radius,
  bladnrs,
  area,
  interpolate = TRUE,
  sheets.location = sheets.location,
  sheets.keep = TRUE,
  sheets.redownload = FALSE
)

Arguments

name
Optional. Give a name of the specified area.

output.dir
Required. Working directory.

AHN
Default 'AHN3'. Set to 'AHN1', 'AHN2', or 'AHN3'.

dem
Default 'DSM'. Choose type of Digital Elevation Model. 'DSM' or 'DTM'. AHN1 only has 'DTM'.

resolution
Default 0.5 meters. Choose resolution of AHN in meters. AHN3 and AHN2 both have 0.5 and 5 meters. AHN1 has 5 and 100 m.

radius
Radius of circle or squared BBOX in meters.

bladnrs
Required. Blad numbers.

area
Required area to be downloaded

interpolate
Default TRUE. Only applicable for AHN2 DTM. It decides if you want the interpolated version of the AHN2 or not.

sheets.location
Optional. Default is the 'AHN_sheets' directory in working directory.. Set directory where all the AHN sheets are stored or will be stored. Always use the correct directory structure and capitalization within the selected directory when pre-existing sheets will be used. Example directory structure: 'AHN_sheets/AHN3/DSM' or 'AHN_sheets/AHN2/DTM'. Only use extracted files in their original name after download.

sheets.keep
Default TRUE. Only applicable if sheets.method is set to TRUE. Set to FALSE if you want to delete the downloaded sheets (structure).

sheets.redownload
Default FALSE. Only applicable if sheets is set to TRUE. Set to TRUE if you want to redownload the sheets (structure)

Source

https://www.pdok.nl/datasets

Value

GeoTIFF of DTM AHN area

Details

Download DTM